Guess My Number Game
Welcome to the Guess My Number Game repository!

Guess My Number is a simple interactive game where players try to guess a randomly generated number within a given range. In this particular implementation, the number to be guessed falls between 0 and 20. The game provides feedback to the player, indicating whether their guess is too high or too low, helping them narrow down the correct answer. The goal is to guess the number in the fewest attempts possible and achieve the highest score.

To experience the game, please visit the Guess My Number Game website. Feel free to explore the code and contribute to the project to enhance its features and functionality. Happy guessing!